pyproject.toml
[tool.poetry]
name = "mvs-cli"
version = "0.1.0"
description = "A CLI for orchestrating the DDI-O MVS workflow"
authors = ["psichael <github@lebaigue.com>"]
readme = "README.md"
packages = [{include = "mvs_harness", from = "."}]

[tool.poetry.dependencies]
python = "~3.10"
requests = ">=2.32.5,<3.0.0"
typer = {extras = ["all"], version = "^0.12.3"}
pydantic = "^2.7.1"
pathspec = "^0.12.1"
jsonschema = "^4.25.1"
click = "==8.1.7"
PyYAML = "^6.0.1"
pyperclip = "^1.8.2"
mistune = "^3.0.2"
pandas = "^2.2.2"
binance = "^0.3.105"
python-binance = "^1.0.35"
scipy = "^1.12.0"
matplotlib = "^3.10.8"
scikit-learn = "^1.5.1"
yfinance = "^1.2.0"

[tool.poetry.dev-dependencies]
pytest = "^8.4.1"
pytest-mock = "^3.14.1"
httpx = "*"

[tool.poetry.scripts]
ddio = "mvs_harness.cli.main:app"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"